Change styling to header for visibility
