--- src/App.css
@@ -1,18 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-body {
-  background-image: url('../public/mushroom.avif');
-  background-size: cover;
-  background-position: inherit;
-  background-repeat: no-repeat;
-  position: relative;
-  margin: 0px;
-  height: 100%;
-} 
-
-html {
-  height: 100%;
-  width: 100%;
-}
