--- src/App.js
@@ -1,16 +1,0 @@
-import './App.css';
-import React, { useState, useEffect } from 'react';
-import Header from './Header/Header';
-
-
-function App() {
-  return (
-    <main className="App">
-      <div className='main-content'>
-        <Header />
-      </div>
-    </main>
-  );
-}
-
-export default App;
