+++ src/Components/Header/Header.css
@@ -1,0 +1,8 @@
+.title-outline {
+  color: rgb(134, 182, 234));
+  text-shadow: 
+    -1px -1px 0 #000,
+    1px -1px 0 #000,
+    -1px 1px 0 #000,
+    1px 1px 0 #000;
+}
